Defines minimum required ChildBenefit Application and routes to it in AppSelector
